% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GimmeMyAI-package.R
\docType{package}
\name{GimmeMyAI-package}
\alias{GimmeMyAI}
\alias{GimmeMyAI-package}
\title{GimmeMyAI: Diagnostics for Multivariate Models}
\description{
Provides a suite of diagnostic tools for multivariate analysis and machine learning. This package implements visualization techniques for dimension reduction methods through correlation circles, score plots, and biplots. It also includes variable importance mapping to facilitate the interpretation of latent structures and the assessment of predictive model performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecamenen/GimmeMyAI}
  \item Report bugs at \url{https://github.com/ecamenen/GimmeMyAI/issues}
}

}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com} (\href{https://orcid.org/0000-0002-2398-0190}{ORCID})

}
\keyword{internal}
